- Recording of units works - updated matlab loadspk.m to handle this - depreciated old loadspike.m function in its own folder
git-svn-id: https://neurorighter.googlecode.com/svn/trunk@375 471701b0-e017-11dd-9759-139c9ab14fc6

+++ NRSpikeSort/NRSpikeSort/SpikeSorter.cs
@@ -213,7 +213,7 @@ namespace NRSpikeSort
                 // Update the newSpikes buffer 
                 for (int j = 0; j < spikesOnChan.Count; ++j)
                 {
-                    spikesOnChan[j].SetUnit(thisChannelModel.classes[j] + thisChannelModel.unitStartIndex + 1);
+                    spikesOnChan[j].SetUnit((Int16)(thisChannelModel.classes[j] + thisChannelModel.unitStartIndex + 1));
                 }
 
             }
